+++ packages/mockyeah-docs/src/layouts/default.js
@@ -7,35 +7,12 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useStaticQuery, graphql, Link } from 'gatsby';
-import isAbsoluteUrl from 'is-absolute-url';
+import { useStaticQuery, graphql } from 'gatsby';
 import { menus } from '../menus';
 import Header from '../components/header';
 import SEO from '../components/seo';
-import Markdown from '../components/Markdown';
+import Menu from './Menu';
 import '../components/layout.css';
-
-const Menu = ({ items }) => (
-  <ul>
-    {items &&
-      items.map(menu => (
-        <li>
-          {menu.url ? (
-            isAbsoluteUrl(menu.url) ? (
-              <a href={menu.url}>{menu.title}</a>
-            ) : (
-              <Link to={menu.url}>
-                <Markdown>{menu.title}</Markdown>
-              </Link>
-            )
-          ) : (
-            <Markdown>{menu.title}</Markdown>
-          )}
-          {menu.items && <Menu items={menu.items} />}
-        </li>
-      ))}
-  </ul>
-);
 
 const Default = ({ children, ...props }) => {
   const data = useStaticQuery(graphql`
@@ -59,21 +36,24 @@ const Default = ({ children, ...props }) => {
       <div
         style={{
           margin: `0 auto`,
-          maxWidth: 960,
+          maxWidth: 1200,
           padding: `0px 1.0875rem 1.45rem`,
           paddingTop: 0
         }}
       >
-        <div style={{ display: 'flex' }}>
+        <div id="page">
+          <main>{children}</main>
           <nav>
             <Menu items={menus.main} />
           </nav>
-          <main>{children}</main>
         </div>
         <footer style={{ textAlign: 'right' }}>
-          <hr />© {new Date().getFullYear()}, Built with
-          {` `}
-          <a href="https://www.gatsbyjs.org">Gatsby</a>
+          <hr />
+          <div>
+            © {new Date().getFullYear()} mockyeah, Built with
+            {` `}
+            <a href="https://www.gatsbyjs.org">Gatsby</a>
+          </div>
         </footer>
       </div>
     </>
