+++ packages/mockyeah/package.json
@@ -63,7 +63,6 @@
     "coveralls": "^3.0.0",
     "dateformat": "^1.0.12",
     "fs-extra": "^7.0.1",
-    "gitbook-plugin-github": "2.0.0",
     "mocha": "^3.1.2",
     "@mockyeah/tools": "^0.24.0-alpha.0",
     "nodemon": "^1.17.5",
